--- pkg/security/ebpf/c/open_filter.h
@@ -1,10 +1,0 @@
-#ifndef _OPEN_FILTER_H_
-#define _OPEN_FILTER_H_
-
-#define BASENAME_FILTER_SIZE 32
-
-struct open_basename_t {
-    char value[BASENAME_FILTER_SIZE];
-};
-
-#endif
