+++ pkg/security/ebpf/c/prebuilt/probe.c
@@ -7,9 +7,13 @@
 
 #include "defs.h"
 #include "buffer_selector.h"
+#include "filters.h"
+#include "approvers.h"
+#include "discarders.h"
+#include "dentry.h"
+#include "exec.h"
 #include "process.h"
 #include "container.h"
-#include "dentry.h"
 #include "overlayfs.h"
 #include "exec.h"
 #include "setattr.h"
@@ -56,6 +60,10 @@ void __attribute__((always_inline)) invalidate_inode(struct pt_regs *ctx, u32 mo
 
         send_event(ctx, EVENT_INVALIDATE_DENTRY, event);
     }
+}
+
+void __attribute__((always_inline)) invalidate_path_key(struct pt_regs *ctx, struct path_key_t *key, int send_invalidate_event) {
+    invalidate_inode(ctx, key->mount_id, key->ino, send_invalidate_event);
 }
 
 __u32 _version SEC("version") = 0xFFFFFFFE;
