+++ pkg/security/ebpf/c/rename.h
@@ -14,12 +14,18 @@ struct rename_event_t {
     u32 padding;
 };
 
+int __attribute__((always_inline)) rename_approvers(struct syscall_cache_t *syscall) {
+    return basename_approver(syscall, syscall->rename.src_dentry, EVENT_RENAME) ||
+           basename_approver(syscall, syscall->rename.target_dentry, EVENT_RENAME);
+}
+
 int __attribute__((always_inline)) trace__sys_rename() {
     struct syscall_cache_t syscall = {
+        .policy = fetch_policy(EVENT_RENAME),
         .type = SYSCALL_RENAME,
     };
 
-    cache_syscall(&syscall, EVENT_RENAME);
+    cache_syscall(&syscall);
 
     return 0;
 }
@@ -51,6 +57,11 @@ int kprobe__vfs_rename(struct pt_regs *ctx) {
     struct dentry *target_dentry = (struct dentry *)PT_REGS_PARM4(ctx);
 
     syscall->rename.src_dentry = src_dentry;
+    syscall->rename.target_dentry = target_dentry;
+
+    if (filter_syscall(syscall, rename_approvers)) {
+        return mark_as_discarded(syscall);
+    }
 
     // use src_dentry as target inode is currently empty and the target file will
     // have the src inode anyway
@@ -70,6 +81,11 @@ int kprobe__vfs_rename(struct pt_regs *ctx) {
         invalidate_inode(ctx, syscall->rename.target_key.mount_id, inode, 1);
     }
 
+    // If we are discarded, we still want to invalidate the inode
+    if (discarded_by_process(syscall->policy.mode, EVENT_RENAME)) {
+        return mark_as_discarded(syscall);
+    }
+
     return 0;
 }
 
@@ -79,6 +95,9 @@ int __attribute__((always_inline)) trace__sys_rename_ret(struct pt_regs *ctx) {
         return 0;
 
     int retval = PT_REGS_RC(ctx);
+    if (IS_UNHANDLED_ERROR(retval)) {
+        return 0;
+    }
 
     u64 inode = get_dentry_ino(syscall->rename.src_dentry);
 
@@ -88,15 +107,9 @@ int __attribute__((always_inline)) trace__sys_rename_ret(struct pt_regs *ctx) {
     }
 
     // invalidate user face inode, so no need to bump the discarder revision in the event
-    invalidate_inode(ctx, syscall->rename.target_key.mount_id, syscall->rename.target_key.ino, 1);
+    invalidate_path_key(ctx, &syscall->rename.target_key, 1);
 
-    // If we are discarded, we still want to invalidate the inode
-    if (discarded_by_process(syscall->policy.mode, EVENT_RENAME) || (IS_UNHANDLED_ERROR(retval))) {
-        return 0;
-    }
-
-    int enabled = is_event_enabled(EVENT_RENAME);
-    if (enabled) {
+    if (!syscall->discarded && is_event_enabled(EVENT_RENAME)) {
         struct rename_event_t event = {
             .syscall.retval = retval,
             .old = {
@@ -121,8 +134,6 @@ int __attribute__((always_inline)) trace__sys_rename_ret(struct pt_regs *ctx) {
 
         send_event(ctx, EVENT_RENAME, event);
     }
-
-    invalidate_inode(ctx, syscall->rename.target_key.mount_id, syscall->rename.target_key.ino, 0);
 
     return 0;
 }
