+++ pkg/security/ebpf/c/rmdir.h
@@ -13,16 +13,23 @@ struct rmdir_event_t {
     u32 padding;
 };
 
+int __attribute__((always_inline)) rmdir_approvers(struct syscall_cache_t *syscall) {
+    return basename_approver(syscall, syscall->rmdir.dentry, EVENT_RMDIR);
+}
+int __attribute__((always_inline)) unlink_approvers(struct syscall_cache_t *syscall);
+
 SYSCALL_KPROBE0(rmdir) {
     struct syscall_cache_t syscall = {
         .type = SYSCALL_RMDIR,
+        .policy = fetch_policy(EVENT_RMDIR),
     };
 
-    cache_syscall(&syscall, EVENT_RMDIR);
+    cache_syscall(&syscall);
 
     return 0;
 }
 
+// security_inode_rmdir is shared between rmdir and unlink syscalls
 SEC("kprobe/security_inode_rmdir")
 int kprobe__security_inode_rmdir(struct pt_regs *ctx) {
     struct syscall_cache_t *syscall = peek_syscall(SYSCALL_RMDIR | SYSCALL_UNLINK);
@@ -50,6 +57,11 @@ int kprobe__security_inode_rmdir(struct pt_regs *ctx) {
             // the mount id of path_key is resolved by kprobe/mnt_want_write. It is already set by the time we reach this probe.
             key = syscall->rmdir.path_key;
 
+            syscall->rmdir.dentry = dentry;
+            if (filter_syscall(syscall, rmdir_approvers)) {
+                return mark_as_discarded(syscall);
+            }
+
             break;
         case SYSCALL_UNLINK:
             event_type = EVENT_UNLINK;
@@ -67,21 +79,23 @@ int kprobe__security_inode_rmdir(struct pt_regs *ctx) {
             // the mount id of path_key is resolved by kprobe/mnt_want_write. It is already set by the time we reach this probe.
             key = syscall->unlink.path_key;
 
+            syscall->unlink.dentry = dentry;
+            syscall->policy = fetch_policy(EVENT_RMDIR);
+            if (filter_syscall(syscall, rmdir_approvers)) {
+                return mark_as_discarded(syscall);
+            }
+
             break;
     }
 
     if (discarded_by_process(syscall->policy.mode, event_type)) {
-        invalidate_inode(ctx, key.mount_id, key.ino, 1);
-
-        return 0;
+        return mark_as_discarded(syscall);
     }
 
     if (dentry != NULL) {
         int ret = resolve_dentry(dentry, key, syscall->policy.mode != NO_FILTER ? event_type : 0);
         if (ret == DENTRY_DISCARDED) {
-            invalidate_inode(ctx, key.mount_id, key.ino, 1);
-
-            pop_syscall(syscall->type);
+            return mark_as_discarded(syscall);
         }
     }
 
@@ -94,14 +108,12 @@ SYSCALL_KRETPROBE(rmdir) {
         return 0;
 
     int retval = PT_REGS_RC(ctx);
-
     if (IS_UNHANDLED_ERROR(retval)) {
-        invalidate_inode(ctx, syscall->rmdir.path_key.mount_id, syscall->rmdir.path_key.ino, 0);
         return 0;
     }
 
-    int enabled = is_event_enabled(EVENT_RMDIR);
-    if (enabled) {
+    int pass_to_userspace = !syscall->discarded && is_event_enabled(EVENT_RMDIR);
+    if (pass_to_userspace) {
         struct rmdir_event_t event = {
             .syscall.retval = retval,
             .file = {
@@ -119,7 +131,7 @@ SYSCALL_KRETPROBE(rmdir) {
         send_event(ctx, EVENT_RMDIR, event);
     }
 
-    invalidate_inode(ctx, syscall->rmdir.path_key.mount_id, syscall->rmdir.path_key.ino, !enabled);
+    invalidate_inode(ctx, syscall->rmdir.path_key.mount_id, syscall->rmdir.path_key.ino, !pass_to_userspace);
 
     return 0;
 }
