+++ pkg/security/ebpf/c/setxattr.h
@@ -13,18 +13,20 @@ struct setxattr_event_t {
 };
 
 int __attribute__((always_inline)) trace__sys_setxattr(const char *xattr_name) {
+    struct policy_t policy = fetch_policy(EVENT_SETXATTR);
+    if (discarded_by_process(policy.mode, EVENT_SETXATTR)) {
+        return 0;
+    }
+
     struct syscall_cache_t syscall = {
         .type = SYSCALL_SETXATTR,
+        .policy = policy,
         .setxattr = {
             .name = xattr_name,
         }
     };
 
-    cache_syscall(&syscall, EVENT_SETXATTR);
-
-    if (discarded_by_process(syscall.policy.mode, EVENT_SETXATTR)) {
-        pop_syscall(SYSCALL_SETXATTR);
-    }
+    cache_syscall(&syscall);
 
     return 0;
 }
@@ -42,14 +44,21 @@ SYSCALL_KPROBE2(fsetxattr, int, fd, const char *, name) {
 }
 
 int __attribute__((always_inline)) trace__sys_removexattr(const char *xattr_name) {
+    struct policy_t policy = fetch_policy(EVENT_REMOVEXATTR);
+    if (discarded_by_process(policy.mode, EVENT_REMOVEXATTR)) {
+        return 0;
+    }
+
     struct syscall_cache_t syscall = {
         .type = SYSCALL_REMOVEXATTR,
+        .policy = policy,
         .setxattr = {
             .name = xattr_name,
         }
     };
 
-    cache_syscall(&syscall, EVENT_REMOVEXATTR);
+    cache_syscall(&syscall);
+
     return 0;
 }
