+++ pkg/security/ebpf/probes/all.go
@@ -62,12 +62,12 @@ func AllMaps() []*manager.Map {
 		{Name: "inode_discarders"},
 		{Name: "pid_discarders"},
 		{Name: "discarder_revisions"},
+		{Name: "basename_approvers"},
 		// Dentry resolver table
 		{Name: "pathnames"},
 		// Snapshot table
 		{Name: "inode_info_cache"},
 		// Open tables
-		{Name: "open_basename_approvers"},
 		{Name: "open_flags_approvers"},
 		// Exec tables
 		{Name: "proc_cache"},
