+++ pkg/security/probe/approvers.go
@@ -8,24 +8,32 @@
 package probe
 
 import (
+	"path"
+
 	"github.com/DataDog/datadog-agent/pkg/security/ebpf"
+	"github.com/DataDog/datadog-agent/pkg/security/model"
+	"github.com/DataDog/datadog-agent/pkg/security/rules"
+	"github.com/DataDog/datadog-agent/pkg/security/secl/eval"
 )
 
+type onApproverHandler func(probe *Probe, approvers rules.Approvers) (activeApprovers, error)
 type activeApprover = activeKFilter
 type activeApprovers = activeKFilters
 
-func approveBasename(tableName string, basename string) (activeApprover, error) {
-	return &mapEntry{
+var allApproversHandlers = make(map[eval.EventType]onApproverHandler)
+
+func approveBasename(tableName string, eventType model.EventType, basename string) (activeApprover, error) {
+	return &mapEventMask{
 		tableName: tableName,
 		key:       basename,
 		tableKey:  ebpf.NewStringMapItem(basename, BasenameFilterSize),
-		value:     ebpf.ZeroUint8MapItem,
+		eventMask: uint64(1 << (eventType - 1)),
 	}, nil
 }
 
-func approveBasenames(tableName string, basenames ...string) (approvers []activeApprover, _ error) {
+func approveBasenames(tableName string, eventType model.EventType, basenames ...string) (approvers []activeApprover, _ error) {
 	for _, basename := range basenames {
-		activeApprover, err := approveBasename(tableName, basename)
+		activeApprover, err := approveBasename(tableName, eventType, basename)
 		if err != nil {
 			return nil, err
 		}
@@ -55,4 +63,80 @@ func setFlagsFilter(tableName string, flags ...int) (activeApprover, error) {
 
 func approveFlags(tableName string, flags ...int) (activeApprover, error) {
 	return setFlagsFilter(tableName, flags...)
+}
+
+func onNewBasenameApprovers(probe *Probe, eventType model.EventType, field string, approvers rules.Approvers) ([]activeApprover, error) {
+	stringValues := func(fvs rules.FilterValues) []string {
+		var values []string
+		for _, v := range fvs {
+			values = append(values, v.Value.(string))
+		}
+		return values
+	}
+
+	prefix := eventType.String()
+	if field != "" {
+		prefix += "." + field
+	}
+
+	var basenameApprovers []activeApprover
+	for field, values := range approvers {
+		switch field {
+		case prefix + ".basename":
+			activeApprovers, err := approveBasenames("basename_approvers", eventType, stringValues(values)...)
+			if err != nil {
+				return nil, err
+			}
+			basenameApprovers = append(basenameApprovers, activeApprovers...)
+
+		case prefix + ".filename":
+			for _, value := range stringValues(values) {
+				basename := path.Base(value)
+				activeApprover, err := approveBasename("basename_approvers", eventType, basename)
+				if err != nil {
+					return nil, err
+				}
+				basenameApprovers = append(basenameApprovers, activeApprover)
+			}
+		}
+	}
+
+	return basenameApprovers, nil
+}
+
+func onNewBasenameApproversWrapper(event model.EventType) onApproverHandler {
+	return func(probe *Probe, approvers rules.Approvers) (activeApprovers, error) {
+		basenameApprovers, err := onNewBasenameApprovers(probe, event, "", approvers)
+		if err != nil {
+			return nil, err
+		}
+		return newActiveKFilters(basenameApprovers...), nil
+	}
+}
+
+func onNewTwoBasenamesApproversWrapper(event model.EventType, field1, field2 string) onApproverHandler {
+	return func(probe *Probe, approvers rules.Approvers) (activeApprovers, error) {
+		basenameApprovers, err := onNewBasenameApprovers(probe, event, field1, approvers)
+		if err != nil {
+			return nil, err
+		}
+		basenameApprovers2, err := onNewBasenameApprovers(probe, event, field2, approvers)
+		if err != nil {
+			return nil, err
+		}
+		basenameApprovers = append(basenameApprovers, basenameApprovers2...)
+		return newActiveKFilters(basenameApprovers...), nil
+	}
+}
+
+func init() {
+	allApproversHandlers["chmod"] = onNewBasenameApproversWrapper(model.FileChmodEventType)
+	allApproversHandlers["chown"] = onNewBasenameApproversWrapper(model.FileChownEventType)
+	allApproversHandlers["link"] = onNewTwoBasenamesApproversWrapper(model.FileLinkEventType, "source", "target")
+	allApproversHandlers["mkdir"] = onNewBasenameApproversWrapper(model.FileMkdirEventType)
+	allApproversHandlers["open"] = openOnNewApprovers
+	allApproversHandlers["rename"] = onNewTwoBasenamesApproversWrapper(model.FileRenameEventType, "old", "new")
+	allApproversHandlers["rmdir"] = onNewBasenameApproversWrapper(model.FileRmdirEventType)
+	allApproversHandlers["unlink"] = onNewBasenameApproversWrapper(model.FileUnlinkEventType)
+	allApproversHandlers["utimes"] = onNewBasenameApproversWrapper(model.FileUtimeEventType)
 }
