+++ pkg/security/probe/open.go
@@ -8,11 +8,11 @@
 package probe
 
 import (
-	"path"
+	"fmt"
 
+	"github.com/DataDog/datadog-agent/pkg/security/model"
 	"github.com/DataDog/datadog-agent/pkg/security/rules"
 	"github.com/DataDog/datadog-agent/pkg/security/secl/eval"
-	"github.com/pkg/errors"
 )
 
 var openCapabilities = Capabilities{
@@ -31,14 +31,6 @@ var openCapabilities = Capabilities{
 }
 
 func openOnNewApprovers(probe *Probe, approvers rules.Approvers) (activeApprovers, error) {
-	stringValues := func(fvs rules.FilterValues) []string {
-		var values []string
-		for _, v := range fvs {
-			values = append(values, v.Value.(string))
-		}
-		return values
-	}
-
 	intValues := func(fvs rules.FilterValues) []int {
 		var values []int
 		for _, v := range fvs {
@@ -47,26 +39,14 @@ func openOnNewApprovers(probe *Probe, approvers rules.Approvers) (activeApprover
 		return values
 	}
 
-	var openApprovers []activeApprover
+	openApprovers, err := onNewBasenameApprovers(probe, model.FileOpenEventType, "", approvers)
+	if err != nil {
+		return nil, err
+	}
+
 	for field, values := range approvers {
 		switch field {
-		case "open.basename":
-			activeApprovers, err := approveBasenames("open_basename_approvers", stringValues(values)...)
-			if err != nil {
-				return nil, err
-			}
-			openApprovers = append(openApprovers, activeApprovers...)
-
-		case "open.filename":
-			for _, value := range stringValues(values) {
-				basename := path.Base(value)
-				activeApprover, err := approveBasename("open_basename_approvers", basename)
-				if err != nil {
-					return nil, err
-				}
-				openApprovers = append(openApprovers, activeApprover)
-			}
-
+		case "open.basename", "open.filename": // already handled by onNewBasenameApprovers
 		case "open.flags":
 			activeApprover, err := approveFlags("open_flags_approvers", intValues(values)...)
 			if err != nil {
@@ -75,7 +55,7 @@ func openOnNewApprovers(probe *Probe, approvers rules.Approvers) (activeApprover
 			openApprovers = append(openApprovers, activeApprover)
 
 		default:
-			return nil, errors.New("field unknown")
+			return nil, fmt.Errorf("unknown field '%s'", field)
 		}
 
 	}
